refactor: AcceptanceTest 의 setUp() 공통부분 추출
